Enable AH_705 Tuning mode
It now seems that the tuner can indeed handle 10 watts in, so the two stage tuning is no longer required. If the IC-705 tuner is set to AH 705, we tune in one stage at up to 10 Watts. If the IC 705 tuner is set to other, we continue to use 2 stage tuning.